requirements.txt
pandas
sentence-transformers
qdrant-client
openai
python-dotenv==1.1.1